backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.31
aiosqlite==0.20.0
asyncpg==0.29.0          # PostgreSQL driver (used in production on Render)
alembic==1.13.1

# Settings & validation
pydantic==2.7.4
pydantic-settings==2.3.4

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── FREE AI / LangChain stack ─────────────────────────────────────────────
langchain==0.2.6
langchain-core==0.2.11
langchain-community==0.2.6

# PRIMARY  → Groq (100% free, no credit card)
# Models: llama3-70b-8192, llama3-8b-8192, mixtral-8x7b-32768, gemma2-9b-it
# Get key: https://console.groq.com
langchain-groq==0.1.6
groq==0.9.0

# FALLBACK → Google Gemini (free tier — 60 req/min on Flash)
# Models: gemini-1.5-flash, gemini-1.5-pro
# Get key: https://aistudio.google.com
langchain-google-genai==1.0.7
google-generativeai==0.7.2

# LangSmith tracing (optional, has a free tier)
# Sign up: https://smith.langchain.com
langsmith==0.1.82

# ── FREE Web search for agents ────────────────────────────────────────────
# Tavily: free tier = 1000 searches/month, no card needed
# Sign up: https://tavily.com
tavily-python==0.3.3

# Task queue
celery==5.4.0
redis==5.0.7

# HTTP client
httpx==0.27.0

# Utils
python-dotenv==1.0.1

